Rename package name, packages in add_could_metadata and add_locale should follow the name of their current folder
