+++ libbeat/processors/add_cloud_metadata/add_cloud_metadata.go
@@ -1,4 +1,4 @@
-package actions
+package add_cloud_metadata
 
 import (
 	"bytes"
