+++ libbeat/processors/add_cloud_metadata/provider_aws_ec2.go
@@ -1,4 +1,4 @@
-package actions
+package add_cloud_metadata
 
 import (
 	"github.com/elastic/beats/libbeat/common"
