+++ libbeat/processors/add_cloud_metadata/provider_aws_ec2_test.go
@@ -1,4 +1,4 @@
-package actions
+package add_cloud_metadata
 
 import (
 	"net/http"
