+++ libbeat/processors/add_cloud_metadata/provider_digital_ocean.go
@@ -1,4 +1,4 @@
-package actions
+package add_cloud_metadata
 
 import (
 	"github.com/elastic/beats/libbeat/common"
