+++ libbeat/processors/add_cloud_metadata/provider_digital_ocean_test.go
@@ -1,4 +1,4 @@
-package actions
+package add_cloud_metadata
 
 import (
 	"net/http"
