+++ libbeat/processors/add_cloud_metadata/provider_tencent_cloud.go
@@ -1,4 +1,4 @@
-package actions
+package add_cloud_metadata
 
 import "github.com/elastic/beats/libbeat/common"
