+++ libbeat/processors/add_locale/add_locale.go
@@ -1,4 +1,4 @@
-package actions
+package add_locale
 
 import (
 	"time"
