+++ libbeat/processors/add_locale/add_locale_test.go
@@ -1,4 +1,4 @@
-package actions
+package add_locale
 
 import (
 	"testing"
